회원 평가 UI변경 및 DB칼럼변경

+++ TeamphonyFinal/src/teamphony/service/facade/MemberService.java
@@ -10,6 +10,6 @@ public interface MemberService {
 	void modifyMember(Member member);
 	void removeMember(String memberId);
 	Member findMemberByMemberId(String memberId);
-	void saveStarPoint(String memberId, int starPoint);
+	void saveStarPoint(String memberId, double starPoint);
 	List<Member> findAllMember();
 }
