+++ TeamphonyFinal/src/teamphony/store/mapper/MemberMapper.java
@@ -12,7 +12,7 @@ public interface MemberMapper {
 	void updateMember(Member member);
 	void deleteMember(String memberId);
 	Member selectMemberByMemberId(String memberId);
-	void insertStarPoint (@Param("memberId") String memberId, @Param("starPoint") int starPoint);
-	List<Integer> getStarPoint(String id);
+	void insertStarPoint (@Param("memberId") String memberId, @Param("starPoint") double starPoint);
+	List<Double> getStarPoint(String id);
 	List<Member> selectAllMember();
 }
